+++ Arena/ArenaLobbyData.cs
@@ -103,6 +103,8 @@ namespace RainMeadow
 
             [OnlineField(group = "arenaSetup")]
             public bool voidMasterEnabled;
+            [OnlineFieldHalf(group = "arenaSetup")]
+            public float voidSpawnLethalityFactor;
 
             [OnlineField(group = "arenaSetup")]
             public int amoebaDuration;
@@ -265,6 +267,7 @@ namespace RainMeadow
 
                 bannedSlugs = new(arena.bannedSlugs);
                 voidMasterEnabled = arena.voidMasterEnabled;
+                voidSpawnLethalityFactor = arena.voidSpawnLethalityFactor;
                 amoebaDuration = arena.amoebaDuration;
                 amoebaControl = arena.amoebaControl;
                 friendlyFire = arena.friendlyFire;
@@ -358,6 +361,8 @@ playerNumberWithTrophiesPerRound;
                 (lobby.gameMode as ArenaOnlineGameMode).bannedSlugs = bannedSlugs;
                 (lobby.gameMode as ArenaOnlineGameMode).piggyBack = piggyBack;
                 (lobby.gameMode as ArenaOnlineGameMode).voidMasterEnabled = voidMasterEnabled;
+                (lobby.gameMode as ArenaOnlineGameMode).voidSpawnLethalityFactor = voidSpawnLethalityFactor;
+
                 (lobby.gameMode as ArenaOnlineGameMode).amoebaDuration = amoebaDuration;
                 (lobby.gameMode as ArenaOnlineGameMode).amoebaControl = amoebaControl;
                 (lobby.gameMode as ArenaOnlineGameMode).friendlyFire = friendlyFire;