+++ Game/RainMeadow.PlayerHooks.cs
@@ -308,6 +308,8 @@ public partial class RainMeadow
         {
             ILCursor c = new(il);
             ILLabel label = null;
+
+            // Proximity & slow down logic ---
             c.GotoNext(MoveType.After, x => x.MatchBneUn(out label));
             c.Emit(OpCodes.Ldarg_0);
             c.Emit(OpCodes.Ldloc, 4);
@@ -325,6 +327,26 @@ public partial class RainMeadow
                 return true;
             });
             c.Emit(OpCodes.Brfalse, label);
+
+            // Lethality
+            c.Index = 0;
+            if (c.TryGotoNext(MoveType.Before,
+                x => x.MatchLdfld<Player>("rippleDeathIntensity"),
+                x => x.MatchLdcR4(out _),
+                x => x.MatchAdd(),
+                x => x.MatchStfld<Player>("rippleDeathIntensity")))
+            {
+                c.Index += 2;
+
+                c.EmitDelegate((float originalIncrement) =>
+                {
+                    if (isArenaMode(out var arena))
+                    {
+                        return originalIncrement * arena.voidSpawnLethalityFactor;
+                    }
+                    return originalIncrement;
+                });
+            }
         }
         catch (Exception ex)
         {